--- a/pandas/tests/io/parser/test_parse_dates.py
+++ b/pandas/tests/io/parser/test_parse_dates.py
@@ -1129,9 +1129,7 @@ def test_parse_dates_empty_string(all_parsers):
     data = "Date,test\n2012-01-01,1\n,2"
     result = parser.read_csv(StringIO(data), parse_dates=["Date"], na_filter=False)
 
-    expected = DataFrame(
-        [[datetime(2012, 1, 1), 1], [pd.NaT, 2]], columns=["Date", "test"]
-    )
+    expected = DataFrame([[datetime(2012, 1, 1), 1], ["", 2]], columns=["Date", "test"])
     tm.assert_frame_equal(result, expected)