@jreback 2020-10-03
why this change? likely to break test parsing generally and would have to see a lot of benefit