@nrebena 2020-10-03
Considering empty string equivalent to NaT as a huge influence on all method that try to cast data into datetimelike (happen in groupby, merge, join, equal, probably others methods)
Globaly, I think we should refrain from **implicit** conversion of string to datetimelike, and this PR goes one step in this direction.
An other issue related to mixing string and datetimelike can be found here #36545 

For information the broken test were the following:
```
pandas/tests/io/parser/test_parse_dates.py::test_parse_dates_empty_string[c_high]
pandas/tests/io/parser/test_parse_dates.py::test_parse_dates_empty_string[c_low]
pandas/tests/io/parser/test_parse_dates.py::test_parse_dates_empty_string[python]
pandas/tests/plotting/test_datetimelike.py::TestTSPlot::test_ts_plot_with_tz['UTC']
pandas/tests/tools/test_to_datetime.py::TestToDatetimeMisc::test_to_datetime_types[True]
pandas/tests/tools/test_to_datetime.py::TestToDatetimeMisc::test_to_datetime_types[False]
pandas/tests/tools/test_to_timedelta.py::TestTimedeltas::test_to_timedelta
```

If empty string should be always considered as NaT, then I'll close this PR.